requirements.txt
torch>=2.1,<3.0
transformers>=4.38,<5.0
tokenizers>=0.15,<0.21

fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
pydantic>=2.0,<3.0

requests>=2.31,<3.0
beautifulsoup4>=4.12,<5.0
python-dotenv>=1.0,<2.0